feat(91.5-87-65): upgrade ModuleManifest to uint64, update FormatConstants/BlockTypeTags
- ModuleManifestField: Value changed from uint to ulong (64-bit manifest)
- ModuleManifestField: Serialize/Deserialize updated to 8-byte ulong LE
- ModuleManifestField: AllModules mask updated to 0x0000_007F_FFFF_FFFFuL (bits 0-38)
- ModuleManifestField: implicit conversions now ulong-based
- FormatConstants: MaxModules=64, DefinedModules=39
- BlockTypeTags: added WALS, ZNSM, OPJR, TRLR tags with FrozenSet entries
- ModuleDefinitions: registry static methods upgraded to accept ulong manifest
- Tier1ModuleVerifier: fixed 1u shift overflow for module IDs >= 32
- Tests updated: buffer 8 bytes, AllModules=39 bits, DefinedModules=39
- VdeComposerTests/VdeCommands: explicit cast ulong -> uint for on-disk compat

+++ DataWarehouse.CLI/Commands/VdeCommands.cs
@@ -91,7 +91,7 @@ public static class VdeCommands
                 ProfileType = VdeProfileType.Custom,
                 Name = "Custom",
                 Description = $"Custom VDE with modules: {modules}",
-                ModuleManifest = manifest.Value,
+                ModuleManifest = (uint)manifest.Value, // cast ulong to uint for VdeCreationProfile (on-disk compat)
                 ModuleConfigLevels = parsedModules.ToDictionary(m => m, _ => (byte)1),
                 BlockSize = blockSize,
                 TotalBlocks = totalBlocks,
