+++ DataWarehouse.Tests/Integration/VdeComposerTests.cs
@@ -85,7 +85,7 @@ public sealed class VdeComposerTests : IDisposable
             ProfileType = VdeProfileType.Custom,
             Name = "SecurityOnly",
             Description = "Test",
-            ModuleManifest = manifest.Value,
+            ModuleManifest = (uint)manifest.Value, // cast ulong to uint for VdeCreationProfile (on-disk compat)
             ModuleConfigLevels = new Dictionary<ModuleId, byte> { [ModuleId.Security] = 1 },
             BlockSize = FormatConstants.DefaultBlockSize,
             TotalBlocks = 1024,
@@ -125,7 +125,7 @@ public sealed class VdeComposerTests : IDisposable
         var path = CreateTempVdePath();
         var profile = VdeCreationProfile.Custom(
             totalBlocks: 1024,
-            manifest: ModuleManifestField.FromModules(ModuleId.Security).Value,
+            manifest: (uint)ModuleManifestField.FromModules(ModuleId.Security).Value,
             configLevels: new Dictionary<ModuleId, byte> { [ModuleId.Security] = 1 },
             blockSize: 8192);
 
@@ -196,7 +196,7 @@ public sealed class VdeComposerTests : IDisposable
     {
         var profile = VdeCreationProfile.Custom(
             1024,
-            ModuleManifestField.FromModules(ModuleId.Security).Value,
+            (uint)ModuleManifestField.FromModules(ModuleId.Security).Value,
             new Dictionary<ModuleId, byte> { [ModuleId.Security] = 1 });
 
         Assert.Equal(VdeProfileType.Custom, profile.ProfileType);
